Reorganized client interface, exposing now services
