+++ examples/create-deploy-app/main.go
@@ -60,17 +60,17 @@ func main() {
 		log.Panic(err)
 	}
 
-	appID, err := client.CreateAppli(appName, appTemplate)
+	appID, err := client.ApplicationService().CreateAppli(appName, appTemplate)
 	if err != nil {
 		log.Panic(err)
 	}
 
-	envID, err := client.GetEnvironmentIDbyName(appID, alien4cloud.DefaultEnvironmentName)
+	envID, err := client.ApplicationService().GetEnvironmentIDbyName(appID, alien4cloud.DefaultEnvironmentName)
 	if err != nil {
 		log.Panic(err)
 	}
 
-	err = client.DeployApplication(appID, envID, locationName)
+	err = client.DeploymentService().DeployApplication(appID, envID, locationName)
 	if err != nil {
 		log.Panic(err)
 	}
@@ -84,7 +84,7 @@ func main() {
 	for !done {
 		time.Sleep(5 * time.Second)
 
-		a4cLogs, nbLogs, err := client.GetLogsOfApplication(appID, envID, filters, logIndex)
+		a4cLogs, nbLogs, err := client.LogService().GetLogsOfApplication(appID, envID, filters, logIndex)
 		if nbLogs > 0 {
 			logIndex = logIndex + nbLogs
 			for idx := 0; idx < nbLogs; idx++ {
@@ -101,7 +101,7 @@ func main() {
 			}
 		}
 
-		status, err := client.GetDeploymentStatus(appID, envID)
+		status, err := client.DeploymentService().GetDeploymentStatus(appID, envID)
 		if err != nil {
 			log.Panic(err)
 		}
@@ -116,7 +116,7 @@ func main() {
 
 	// On succesful deployment print output variable if any
 	if deploymentStatus == alien4cloud.ApplicationDeployed {
-		nodeAttrOutputs, err := client.GetOutputAttributes(appID, envID)
+		nodeAttrOutputs, err := client.DeploymentService().GetOutputAttributes(appID, envID)
 		if err != nil {
 			log.Panic(err)
 		}
@@ -124,7 +124,7 @@ func main() {
 		if len(nodeAttrOutputs) > 0 {
 			fmt.Println("\nOutputs:")
 			for nodeName, attrs := range nodeAttrOutputs {
-				attrValues, err := client.GetAttributesValue(appID, envID, nodeName, attrs)
+				attrValues, err := client.DeploymentService().GetAttributesValue(appID, envID, nodeName, attrs)
 				if err != nil {
 					log.Panic(err)
 				}
